PortfolioManagementSystem/requirements/base.txt
Django
django-mptt
django-rest-framework
Pillow
markdown
psycopg2
